I've included the original test documentation/spec within the repository.
I've also started with defining the types for Users and Groups with
mappings for JSON marshalling.

After looking at my options, it looks like the sleepy micro-framework is
a good start for something this simple.  It's written by the current
maintainer of Python-Flask.  I considered just implementing the entire
thing from scratch using the net/http core library, but dealing with
routes is a bit troublesome and I didn't want to deal with gorilla/mux.