Rather than writing a bunch of helper functions to try to deal with the constraint of having my User objects contain a Groups object and my Group objects contain a Users object, I instead ceded to using a slice of strings and writing much less complicated and awful helper functions to generate GroupsWithMembers on the fly when requested.  Therefore, after this change, User objects contain a Groups []string which gets parsed as needed for Group related functions.  A Group object only contains a GroupName string, and is the only object stored. GroupsWithMembers are discarded after use.
I'm not sure that this is the best way to approach this, but short of
implementing a SQL database to back this it was the simplest way to
enforce tight consistency between my data types while providing foreign
constraints and uniqueness.  I may end up refactoring the whole thing
again but meh.

As an addition, I also added an error handler for NotFound (404) and
started on the group HTTP handlers.  The user HTTP handlers now check
groups as well and respond appropriately.  My handlers are kind of
getting huge but guh.  This whole thing is way gnarlier than I'd like.
I guess this is why people use frameworks.